blp/requirements.txt
discord.py
random
math
sys
traceback
json
os
dotenv
asyncio
datetime
platform
colorama
PIL
io